# Faster gather in common cases