--- a/mlx/backend/common/compiled.cpp
+++ b/mlx/backend/common/compiled.cpp
@@ -1,0 +1,10 @@
+// Copyright © 2023-2024 Apple Inc.
+
+#include "mlx/allocator.h"
+#include "mlx/primitives.h"
+
+namespace mlx::core {
+void Compiled::eval(const std::vector<array>& inputs, std::vector<array>& out) {
+}
+
+} // namespace mlx::core