… benchmarks/python/gather_bench.py
… mlx/backend/accelerate/primitives.cpp
… mlx/backend/common/CMakeLists.txt
… mlx/backend/common/default_primitives.cpp
… mlx/backend/metal/CMakeLists.txt
… mlx/backend/no_metal/primitives.cpp
… mlx/compile.h
… mlx/primitives.h
… mlx/transforms.h
… python/tests/test_compile.py
… tests/compile_tests.cpp