## Proposed changes

- Add a 0 and 1D specialization for the indices dimension
- Use a 2D grid instead of 1D grid to simplify and speedup indexing into the src array
- Add a gather benchmark inspired from #518 

Closes #518 

## Benchmarks

### Before:

```
====================
X (100, 64), Indices (1000000,)
MLX: 28.516ms
PyTorch: 12.662ms
====================
X (100, 1024), Indices (100000,)
MLX: 39.227ms
PyTorch: 20.165ms
====================
X (4, 1000000), Indices ()
MLX: 0.326ms
PyTorch: 0.171ms
```


### After:
```
====================
X (100, 64), Indices (1000000,)
MLX: 3.155ms
PyTorch: 12.978ms
====================
X (100, 1024), Indices (100000,)
MLX: 4.875ms
PyTorch: 20.573ms
====================
X (4, 1000000), Indices ()
MLX: 0.288ms
PyTorch: 0.185ms
```